% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{catable}
\alias{catable}
\title{Category table summarizer}
\usage{
catable(
  data,
  categories = c(quantile(data, c(0.01, 0.1, 0.5, 0.9, 0.99), na.rm = TRUE)),
  cumulative = FALSE,
  na.rm = TRUE,
  digits = 3
)
}
\arguments{
\item{data}{Numeric vector.}

\item{categories}{Numeric vector of category boundaries.}

\item{cumulative}{Logical, whether cumulative counts should be returned.}

\item{na.rm}{Logical, remove NAs before computation.}

\item{digits}{Number of digits to round results.}
}
\value{
A 2-row matrix summarizing counts and proportions.
}
\description{
Produce a table summarizing how many values fall within specified cut
categories, optionally cumulative.
}
