% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{detect_species_from_symbol}
\alias{detect_species_from_symbol}
\title{Detect species (human or mouse) from gene symbols}
\usage{
detect_species_from_symbol(gene)
}
\arguments{
\item{gene}{Character vector of gene symbols.}
}
\value{
"human", "mouse", or "unknown".
}
\description{
Detect species (human or mouse) from gene symbols
}
