% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{.parse_ratio_to_numeric}
\alias{.parse_ratio_to_numeric}
\title{Dot plot for GO enrichment (style 1)}
\usage{
.parse_ratio_to_numeric(x)
}
\arguments{
\item{GO_df}{A GO enrichment result data frame.}

\item{topn}{Number of categories to show.}

\item{label_format}{Width for wrapped text labels.}

\item{group_by}{Metric used to choose topN terms. One of
\code{"-log10(p.adjust)"}, \code{"GeneRatio"}, \code{"RichFactor"},
\code{"Count"}. Default is \code{"-log10(p.adjust)"}.}
}
\value{
A ggplot object.
}
\description{
This function takes a GO enrichment result (e.g. clusterProfiler output)
and visualizes the top enriched categories. Top terms can be selected
by different metrics.
}
