% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EasyProtein-package.R
\docType{package}
\name{EasyProtein}
\alias{EasyProtein-package}
\alias{EasyProtein}
\title{EasyProtein: A comprehensive proteomics analysis toolkit}
\description{
This package provides QC, normalization, DE analysis, and visualization
for proteomics workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yuanlizhanshi/EasyProtein}
  \item Report bugs at \url{https://github.com/yuanlizhanshi/EasyProtein/issues}
}

}
\author{
\strong{Maintainer}: Yunhui Kong \email{kongyunhui1@gamil.com}

}
