% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{fix_duplicate_protein_ids}
\alias{fix_duplicate_protein_ids}
\title{Fix duplicated protein IDs}
\usage{
fix_duplicate_protein_ids(df, id_col = "Protein.Ids")
}
\arguments{
\item{df}{A data.frame containing protein information.}

\item{id_col}{Column name containing protein identifiers.}
}
\value{
A data.frame with fixed unique IDs.
}
\description{
Detect and automatically rename duplicated IDs in a proteomics matrix.
}
