% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geom_mean}
\alias{geom_mean}
\title{Geometric mean}
\usage{
geom_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector.}

\item{na.rm}{Whether to remove NA values.}
}
\value{
A numeric value representing the geometric mean.
}
\description{
Compute geometric mean, optionally removing NA values.
}
