% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_batch_DEGs}
\alias{get_batch_DEGs}
\title{Run pairwise DEG comparisons for all conditions except a reference}
\usage{
get_batch_DEGs(se, compare_col = "condition", start_col = 1)
}
\arguments{
\item{se}{A SummarizedExperiment object.}

\item{compare_col}{Column in colData(se) representing conditions.}

\item{start_col}{Index specifying which condition to use as reference.}
}
\value{
A named list of DEG data.frames.
}
\description{
Run pairwise DEG comparisons for all conditions except a reference
}
