% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_turning_point}
\alias{get_turning_point}
\title{Count turning points in vector}
\usage{
get_turning_point(vec, fluctuate = 0.05)
}
\arguments{
\item{vec}{Numeric vector.}

\item{fluctuate}{Allowed fluctuation adjustment.}
}
\value{
Number of turning points.
}
\description{
Count turning points in vector
}
