% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_easyprotein.R
\name{launch_EasyProtein}
\alias{launch_EasyProtein}
\title{Launch EasyProtein Shiny App}
\usage{
launch_EasyProtein(...)
}
\arguments{
\item{...}{Passed to shiny::runApp().}
}
\description{
Launch the EasyProtein interactive Shiny application.
}
