% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_mtx}
\alias{scale_mtx}
\title{Scale matrix by row}
\usage{
scale_mtx(mtx)
}
\arguments{
\item{mtx}{A numeric matrix.}
}
\value{
A scaled matrix with the same dimensions.
}
\description{
Scale each row of a numeric matrix to mean 0 and sd 1.
}
