% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{se2internstiy}
\alias{se2internstiy}
\title{Convert SummarizedExperiment to intensity data.frame}
\usage{
se2internstiy(se)
}
\arguments{
\item{se}{A SummarizedExperiment object.}
}
\value{
A data.frame containing rowData and intensity assay.
}
\description{
Convert SummarizedExperiment to intensity data.frame
}
