% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{se2raw}
\alias{se2raw}
\title{Convert SummarizedExperiment to raw intensity data.frame}
\usage{
se2raw(se)
}
\arguments{
\item{se}{A SummarizedExperiment object.}
}
\value{
A data.frame containing rowData and raw_intensity assay.
}
\description{
Convert SummarizedExperiment to raw intensity data.frame
}
