% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize_se_by_coldata}
\alias{summarize_se_by_coldata}
\title{Summarize assay values by groups in colData(se)}
\usage{
summarize_se_by_coldata(
  se,
  group_by,
  assay_name = "conc",
  method = c("mean", "median"),
  na.rm = TRUE
)
}
\arguments{
\item{se}{SummarizedExperiment object.}

\item{group_by}{Column defining groups.}

\item{assay_name}{Assay to summarize.}

\item{method}{"mean" or "median".}

\item{na.rm}{Logical.}
}
\value{
A matrix of summarized expression.
}
\description{
Summarize assay values by groups in colData(se)
}
