% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_GO_dot2}
\alias{plot_GO_dot2}
\title{Dot plot for GO enrichment (style 2)}
\usage{
plot_GO_dot2(GO_df, topn = 10, label_format = 30)
}
\arguments{
\item{GO_df}{A GO enrichment result data frame.}

\item{topn}{Number of GO terms.}

\item{label_format}{Width for wrapped labels.}
}
\value{
A ggplot object.
}
\description{
This visualization computes GeneRatio and displays enriched GO terms
using a colored dot plot.
}
