% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_GO_dot3}
\alias{plot_GO_dot3}
\title{Dot plot for GO/Pathway output (style 3, generic)}
\usage{
plot_GO_dot3(GO_df, topn = 10, label_format = 30)
}
\arguments{
\item{GO_df}{A generic enrichment result table.}

\item{topn}{Number of terms.}

\item{label_format}{Width for wrapping labels.}
}
\value{
A ggplot object.
}
\description{
A flexible dot plot for enrichment tables containing terms such as:
"term description", "enrichment score", "false discovery rate".
}
