% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotCV_density}
\alias{plotCV_density}
\title{Plot CV distribution across conditions}
\usage{
plotCV_density(se)
}
\arguments{
\item{se}{A SummarizedExperiment object.}
}
\value{
A ggplot object or NULL if no replicates.
}
\description{
Plot CV distribution across conditions
}
