% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotSE_density}
\alias{plotSE_density}
\title{Plot intensity distribution for each sample}
\usage{
plotSE_density(se)
}
\arguments{
\item{se}{A SummarizedExperiment object.}
}
\value{
A ggplot violin plot.
}
\description{
This function draws violin + boxplot of log2 intensity for each sample
from a SummarizedExperiment object.
}
