% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotSE_protein_number}
\alias{plotSE_protein_number}
\title{Plot number of detected proteins per sample}
\usage{
plotSE_protein_number(se)
}
\arguments{
\item{se}{A SummarizedExperiment object.}
}
\value{
A ggplot object.
}
\description{
Plot number of detected proteins per sample
}
