% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{rawdata2se}
\alias{rawdata2se}
\title{Import raw data and construct SummarizedExperiment for QC}
\usage{
rawdata2se(
  exp_file,
  obs_col = "Protein.Ids",
  enable_detect_outlier_gene = FALSE,
  fc_threshold = 5,
  valid_group_cutoff = -1,
  frac_NA_threshold = 0.5,
  stable_group_cutoff = -1,
  cv_threshod = 0.5
)
}
\arguments{
\item{exp_file}{Path to raw intensity file.}

\item{obs_col}{Column containing protein identifiers.}

\item{enable_detect_outlier_gene}{Logical, perform FC-based outlier removal.}

\item{fc_threshold}{Fold-change threshold for detecting outliers.}

\item{valid_group_cutoff}{Minimum number of valid groups after NA filtering.}

\item{frac_NA_threshold}{NA fraction threshold per group.}

\item{stable_group_cutoff}{Minimum number of groups with stable CV.}

\item{cv_threshod}{CV threshold for determining stable genes.}
}
\value{
A list containing:
\describe{
  \item{se}{A \code{SummarizedExperiment} object}
  \item{un_stable_gene}{Genes removed due to unstable CV}
  \item{missing_gene_df}{Genes removed due to excessive missingness}
}
}
\description{
Read MaxQuant-like protein intensity matrix, perform QC filtering,
remove missing genes, impute missing values, filter unstable proteins,
and return a \code{SummarizedExperiment} object containing raw,
imputed, CPM, and z-scaled assays.
}
