% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{se2conc}
\alias{se2conc}
\title{Convert SummarizedExperiment to concentration (CPM) data.frame}
\usage{
se2conc(se)
}
\arguments{
\item{se}{A SummarizedExperiment object.}
}
\value{
A data.frame containing rowData and conc assay.
}
\description{
Convert SummarizedExperiment to concentration (CPM) data.frame
}
