% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{se2gene_group}
\alias{se2gene_group}
\title{Assign genes into dynamic groups based on expression, correlation, and CV}
\usage{
se2gene_group(
  se,
  assay = "conc",
  group_by = "condition",
  time_threhold = 0.75,
  min_expresion_threhold = 1,
  CV_with_time_threhold = 0.1,
  padj_threhold = 0.05
)
}
\arguments{
\item{se}{A SummarizedExperiment object.}

\item{assay}{Assay name to use (default "conc").}

\item{group_by}{Column in colData(se) defining sample ordering.}

\item{time_threhold}{thresholds for up/down classification.}

\item{min_expresion_threhold}{Minimum mean expression.}

\item{CV_with_time_threhold}{CV threshold used to define stable genes.}

\item{padj_threhold}{Adjusted p-value threshold for DEGs.}
}
\value{
Updated SummarizedExperiment with gene grouping added to rowData.
}
\description{
This function identifies gene behavior patterns (Up, Down, Stable,
fluctuated, Low expression) based on:
- time-course correlation
- differential expression
- mean expression
- coefficient of variation (CV)
}
