% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{se2scale}
\alias{se2scale}
\title{Convert SummarizedExperiment to z-scaled CPM data.frame}
\usage{
se2scale(se)
}
\arguments{
\item{se}{A SummarizedExperiment object.}
}
\value{
A data.frame containing rowData and zscale assay.
}
\description{
Convert SummarizedExperiment to z-scaled CPM data.frame
}
